add under_attack map to the Board class and methods for Wazir Piece
The under_attack is a map that tells, which square is under attack by
which piece. Each piece will gets new methods validate_attack() and
return_attack()

The return_attack() returns all relative squares that can be attacked
by the piece depending on its amount of movement_points (movement_pts.
part not fully developed yet)

The validate_attack() method gets all the relative squares' passablity
and all pieces that migth be there. The point of the method is to cut
off any attack squares that can't be reached, e.g. if the same color
piece stands in front of the piece the validate_attack() method will
not return any forward-facing attack squares, except for the piece
standing in front (because it will be defended)